@@ -4,14 +4,16 @@
   "private": true,
   "homepage": "http://jonparker.github.io/movies-frontend",
   "dependencies": {
-    "@material-ui/core": "^4.6.1",
-    "@material-ui/icons": "^4.5.1",
-    "@material-ui/lab": "^4.0.0-alpha.32",
-    "prop-types": "^15.7.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "^4.0.3",
-    "typeface-roboto": "0.0.75"
+    "@emotion/css": "^11.9.0",
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
+    "@mui/icons-material": "^5.6.2",
+    "@mui/material": "^5.7.0",
+    "prop-types": "^15.8.1",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-scripts": "^5.0.1",
+    "typeface-roboto": "1.1.13"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -40,21 +42,21 @@
     ]
   },
   "devDependencies": {
-    "@types/jest": "24.0.22",
-    "@types/node": "12.12.7",
-    "@types/prop-types": "^15.7.3",
-    "@types/react": "^17.0.6",
-    "@types/react-dom": "^17.0.5",
-    "@typescript-eslint/eslint-plugin": "^4.24.0",
-    "@typescript-eslint/parser": "^4.24.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-jest": "^24.3.6",
+    "@types/jest": "27.5.1",
+    "@types/node": "17.0.33",
+    "@types/prop-types": "^15.7.5",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.4",
+    "@typescript-eslint/eslint-plugin": "^5.23.0",
+    "@typescript-eslint/parser": "^5.23.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-jest": "^26.2.2",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-react": "^7.23.2",
-    "gh-pages": "^2.1.1",
-    "parcel": "^2.0.0-beta.3.1",
-    "prettier": "^2.3.0",
+    "eslint-plugin-react": "^7.29.4",
+    "gh-pages": "^4.0.0",
+    "parcel": "^2.5.0",
+    "prettier": "^2.6.2",
     "tslint": "^6.1.3",
-    "typescript": "^4.2.4"
+    "typescript": "^4.6.4"
   }
 }
