@@ -1,43 +1,40 @@
-import React from "react";
-import { makeStyles, createStyles, Theme } from "@material-ui/core/styles";
-import Grid from "@material-ui/core/Grid";
-import Paper from "@material-ui/core/Paper";
-import IMovieProps from "./IMovieProps";
-import MovieCard from "./MovieCard";
+/* eslint-disable react/react-in-jsx-scope -- Unaware of jsxImportSource */
+/** @jsxImportSource @emotion/react */
 
-const useStyles = makeStyles((theme: Theme) =>
-  createStyles({
-    root: {
-      flexGrow: 1,
-    },
-    paper: {
-      padding: theme.spacing(2),
-      margin: "auto",
-    },
-    grid: {
-      gap: "1rem",
-    },
-  })
-);
+import React from "react"
+import { css } from "@emotion/react"
+import Grid from "@mui/material/Grid"
+import Paper from "@mui/material/Paper"
+import IMovieProps from "./IMovieProps"
+import MovieCard from "./MovieCard"
 
-const Movies: React.FC<{ movies: Array<IMovieProps> }> = ({
-  movies,
-}: {
-  movies: IMovieProps[];
-}) => {
-  const classes = useStyles();
-
+const Movies = ({ movies }: { movies: Array<IMovieProps> }) => {
   return (
-    <div className={classes.root}>
-      <Paper square className={classes.paper}>
-        <Grid container spacing={3} className={classes.grid}>
+    <div
+      css={css`
+        flexgrow: 1;
+      `}
+    >
+      <Paper
+        square
+        css={css`
+          padding: 2rem;
+        `}
+      >
+        <Grid
+          container
+          spacing={3}
+          css={css`
+            gap: 1rem;
+          `}
+        >
           {movies.map((movie) => (
             <MovieCard movie={movie} key={movie.id} />
           ))}
         </Grid>
       </Paper>
     </div>
-  );
-};
+  )
+}
 
-export default Movies;
+export default Movies
